+++ packages/basic/src/widgets/image-widget.ts
@@ -1,7 +1,6 @@
 import {
   Widget,
   WidgetOptions,
-  $sourcesLoaded,
   AsyncWidget,
   AsyncWidgetResponse,
 } from '@newcar/core'