+++ packages/core/src/index.ts
@@ -6,6 +6,5 @@ export * from './scene'
 export * from './widget'
 export * from './asyncWidget'
 
-export * from './apiPreload'
 export * from './apiChangeProperty'
 export * from './apiChangeStyle'