+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @newcar/basic
 
+## 0.10.0
+
+### Minor Changes
+
+- This version add Recorder package and fix known bug.
+
+### Patch Changes
+
+- Updated dependencies
+  - @newcar/core@0.10.0
+  - @newcar/utils@0.10.0
+
 ## 0.9.0
 
 ### Minor Changes