+++ packages/newcar/CHANGELOG.md
@@ -1,5 +1,19 @@
 # newcar
 
+## 0.10.0
+
+### Minor Changes
+
+- This version add Recorder package and fix known bug.
+
+### Patch Changes
+
+- Updated dependencies
+  - @newcar/recorder@0.10.0
+  - @newcar/basic@0.10.0
+  - @newcar/core@0.10.0
+  - @newcar/utils@0.10.0
+
 ## 0.9.0
 
 ### Minor Changes