+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @newcar/utils
 
+## 0.10.0
+
+### Minor Changes
+
+- This version add Recorder package and fix known bug.
+
 ## 0.9.0
 
 ### Minor Changes